feat(projects): pre generate project resouces
