+++ package.json
@@ -33,7 +33,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "generate-resources": "node ./scripts/generate-resources.js"
   },
   "eslintConfig": {
     "extends": [
