--- src/projects/maintainerProjects.js
@@ -1,11 +1,0 @@
-import { CAMUNDA_MODELER_TAG, NODE_TAG, PERSONAL_TAG } from "./tags";
-
-const maintainerProjects = [
-    {
-        name: "Camunda Modeler Provisioner",
-        github: "https://github.com/AlexanderSkrock/camunda-modeler-provisioner",
-        tags: [ PERSONAL_TAG, CAMUNDA_MODELER_TAG, NODE_TAG ]
-    }
-];
-
-export default maintainerProjects;
