--- src/projects/tags.js
@@ -1,14 +1,0 @@
-// Scope of project
-export const PERSONAL_TAG = "Personal";
-
-// Languages
-export const JSON_SCHEMA_TAG = "JSON Schema";
-export const JAVA_TAG = "Java";
-export const JS_TAG = "JavaScript";
-
-// Framework & Technologies
-export const NODE_TAG = "Node.js";
-export const REACT_TAG = "React";
-export const SPRING_BOOT_TAG = "Spring Boot";
-export const CAMUNDA_7_TAG = "Camunda 7";
-export const CAMUNDA_MODELER_TAG = "Camunda Modeler";
